Refactor message handling, add file transport
The message handling is refactored to clean up bot.on('message',
...). The quip is picked only when needed. Command actions (e.g.,
!learn, !dump) are abstracted to appropriate functions.

Created command action functions:
  * delayedMessage: sends a message to chat after fixed delay
  * logDump: dumps log to chat
  * learnTarget: switches to new target
  * pickQuip: reads message.log and picks a quip.

A new logger transport that writes to file is created. The log
is currently reused between executions of the bot, so it will
dump logs from the previous run as well.

To make the constants more readable, they are now attributes of
bot and can be modified. Initialization of bot's extra
attributes are grouped together, but these can be put in a new
object.