Add !set, bot init style, add disconnect case
A dangerous !set command is implemented. This allows the user to
set any configurable property of the bot.

The configurable properties of the bot (i.e., the custom
properties that we defined and attached to the Discord.Client
instance) are now wrapped in a settableProperties object from
which we copy each attribute over to `bot`.

If the Discord bot disconnects (or fails to connect) it now
logs an error to the logger.log and console.