Feature: cli debug and other commands introduced (#9)
* Refactor message handling, add file transport

The message handling is refactored to clean up bot.on('message',
...). The quip is picked only when needed. Command actions (e.g.,
!learn, !dump) are abstracted to appropriate functions.

Created command action functions:
  * delayedMessage: sends a message to chat after fixed delay
  * logDump: dumps log to chat
  * learnTarget: switches to new target
  * pickQuip: reads message.log and picks a quip.

A new logger transport that writes to file is created. The log
is currently reused between executions of the bot, so it will
dump logs from the previous run as well.

To make the constants more readable, they are now attributes of
bot and can be modified. Initialization of bot's extra
attributes are grouped together, but these can be put in a new
object.

* Add pretty formatting, change logger setup syntax

* Add !env command, wrote down other commands

* Clean bot.on('message'), improve logging

* Add flag parsing, add flags to !log

* Add documentation to logDump function

* Add !help, log file uploading commands

* Add targetPlainString(), minor fix in delayedMessage signature

* Linting

* Add !set, bot init style, add disconnect case

A dangerous !set command is implemented. This allows the user to
set any configurable property of the bot.

The configurable properties of the bot (i.e., the custom
properties that we defined and attached to the Discord.Client
instance) are now wrapped in a settableProperties object from
which we copy each attribute over to `bot`.

If the Discord bot disconnects (or fails to connect) it now
logs an error to the logger.log and console.

* Update package-lock.json and package.json

Only these files are updated. Had a problem with reverting my
discord.io library to a version that gets rejected by the Discord
servers; see https://github.com/izy521/discord.io/issues/234.

Can revert these if needed.

* Remove extraneous log

* Add !dance, interactive debugging

!dance reacts with a link to a dancing pepe. Nothing is stopping
a broken link from being posted, so we better watch out.

Interactive debugging is made available via !db --it, and can be
switched off via !db --off. The debugging currently listens on all
channels, which is a little annoying.

* Add goodbot, badbot, yikes reactions; add env process architecture

* Remove extra transport, sort !help